fix: moved to long the default store size

+++ src/Carter.Cache/Stores/MemoryStore.cs
@@ -18,7 +18,7 @@ namespace Carter.Cache.Stores
             sizeLimit = options.SizeLimit ?? 0;
         }
 
-        public DefaultMemoryStore(int maxSize) : this(new MemoryCacheOptions() { SizeLimit = maxSize })
+        public DefaultMemoryStore(long maxSize) : this(new MemoryCacheOptions() { SizeLimit = maxSize })
         {
         }
